Desenvolvimento do Projeto

+++ Projeto Integrador I/Loja de Jogos/deletargeneroj.py
@@ -2,19 +2,34 @@ import mysql.connector
 from connector import host, user, password, database
 
 conexao = mysql.connector.connect(
-
-    host = host,
-    user = user,
-    password= password,  
-    database= database,  
+    host=host,
+    user=user,
+    password=password,
+    database=database
 )
 
-cursor = conexao.cursor()
-
 def deletar_generoj():
-    cursor = conexao.cursor()
-    idgeneroj = input("\nDigite o ID do genero a ser deletado: ")
-    sql = f"DELETE FROM generosjogo WHERE idgeneroj = '{idgeneroj}'"
-    cursor.execute(sql)
-    conexao.commit()
-    print("\nGenero deletado com sucesso!")
+    while True:
+        try:
+            idgeneroj = input("\nDigite o ID do gênero a ser deletado: ")
+
+            if not idgeneroj.isdigit():
+                print("ID inválido. Digite novamente.")
+                continue
+
+            cursor = conexao.cursor()
+            sql = f"DELETE FROM generosjogo WHERE idgeneroj = '{idgeneroj}'"
+            cursor.execute(sql)
+            conexao.commit()
+            print("\nGênero deletado com sucesso!")
+            break
+
+        except mysql.connector.Error as error:
+            print(f"Erro ao deletar o gênero: {error}")
+            continue
+
+        break
+
+# deletar_generoj()
+
+# conexao.close()
